+++ collection/templates/collection-spec.js
@@ -13,21 +13,24 @@
             this.<%= _.classify(name) %>Collection = new <%= _.classify(name) %>();
         });
 
+        it('Should run a few assertions', function(){
+
+        });
+
     });
 
 });<% } else if (jsOption === 'Browserify') { %>'use strict';
 
 var <%= _.classify(name) %> = require('../../dev/scripts/collections/<%= _.slugify(name) %>.js');
 
-describe('just checking', function() {
+describe('<%= _.classify(name) %> Collection', function () {
 
-    describe('<%= _.classify(name) %> Collection', function () {
+    beforeEach(function () {
+        this.<%= _.classify(name) %>Collection = new <%= _.classify(name) %>();
+    });
 
-        beforeEach(function () {
-            this.<%= _.classify(name) %>Collection = new <%= _.classify(name) %>();
-        });
+    it('Should run a few assertions', function(){
 
     });
 
-});
-<% } %>
+});<% } %>