+++ script/templates/script-spec.js
@@ -7,14 +7,10 @@
 
     var <%= _.classify(name) %> = require('modules/<%= _.slugify(name) %>');
 
-    describe('just checking', function() {
+    describe('Give it some context', function() {
 
-        it('works for <%= _.camelize(name) %>', function() {
-            var msg = '<%= _.camelize(name) %> loaded!';
+        it('Should run a few assertions', function() {
 
-            var message = <%= _.classify(name) %>.init(msg);
-
-            expect(message).toMatch(/initialized/);
         });
 
     });
@@ -23,28 +19,19 @@
 
 var <%= _.classify(name) %> = require('../../../dev/scripts/modules/<%= _.slugify(name) %>.js');
 
-describe('just checking', function() {
+describe('Give it some context', function() {
 
-    it('works for <%= _.camelize(name) %>', function() {
-        var msg = '<%= _.camelize(name) %> loaded!';
+    it('Should run a few assertions', function() {
 
-        var message = <%= _.classify(name) %>.init(msg);
-
-        expect(message).toMatch(/initialized/);
     });
 
 });
 <% } else { %>'use strict';
 
-describe('just checking <%= _.camelize(name) %>', function() {
+describe('Give it some context', function() {
 
-    it('works for <%= _.camelize(name) %>', function() {
-        var msg = '<%= _.camelize(name) %> is working!';
+    it('Should run a few assertions', function() {
 
-        var message = <%= _.camelize(name) %>.init(msg);
-
-        expect(message).toMatch(/initialized/);
     });
 
-});
-<% } %>
+});<% } %>