+++ view/templates/view-spec.js
@@ -14,21 +14,24 @@ define(function(require) {
             this.<%= _.classify(name) %>View = new <%= _.classify(name) %>();
         });
 
+        it('Should run a few assertions', function(){
+
+        });
+
     });
 
 });<% } else if (jsOption === 'Browserify') { %>'use strict';
 
 var <%= _.classify(name) %> = require('../../dev/scripts/views/<%= _.slugify(name) %>.js');
 
-describe('just checking', function() {
+describe('<%= _.classify(name) %> View', function () {
 
-    describe('<%= _.classify(name) %> View', function () {
+    beforeEach(function () {
+        this.<%= _.classify(name) %>View = new <%= _.classify(name) %>();
+    });
 
-        beforeEach(function () {
-            this.<%= _.classify(name) %>View = new <%= _.classify(name) %>();
-        });
+    it('Should run a few assertions', function(){
 
     });
 
-});
-<% } %>
+});<% } %>